Complete Gameboard
Added the gameboard to the page. There is an error with the sizing of
the boxes within the board. The height does not want to resize.

The sign in, sign out, sign up and change password still need to be
adjusted to hide when they do not need to be in use.