Added features
Show/hide features for sign out and password change to not show on the
sign in and sign up page. There is no sign in or sign up showing when
one is currently logged in, only the sign out and change password shows.
Added legends for the sign in and sign up in the html.