Game board
The gameboard now hides before log in and after the game is over. This
prevents the player from continuing to click on the board.
New Game shows the game board but does NOT reset it (yet).

Working on the Tie Function, in addition to getting the game board to
reset when clicking New Game